requirements.txt
# Telegram Bot Framework
python-telegram-bot==20.8
aiohttp==3.9.3
asyncio-compat==0.1.2

# Logging and utilities
python-dotenv==1.0.0
requests==2.31.0

# For web scraping (if needed)
beautifulsoup4==4.12.2
lxml==4.9.3

# For better async support
aiosignal==1.3.1
